// packages/backend-lib/src/twentyCrm/metaClient/model/field.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Twenty Api
 * This is a **Twenty REST/API** playground based on the **OpenAPI 3.1 specification**.
 *
 * The version of the OpenAPI document: v0.1
 * Contact: felix@twenty.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FieldOptionsInner } from './field-options-inner';

/**
 * A field
 * @export
 * @interface Field
 */
export interface Field {
    /**
     * 
     * @type {string}
     * @memberof Field
     */
    'type': FieldTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof Field
     */
    'objectMetadataId': string;
    /**
     * 
     * @type {string}
     * @memberof Field
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof Field
     */
    'label': string;
    /**
     * 
     * @type {string}
     * @memberof Field
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof Field
     */
    'icon'?: string;
    /**
     * 
     * @type {any}
     * @memberof Field
     */
    'defaultValue'?: any;
    /**
     * 
     * @type {boolean}
     * @memberof Field
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {object}
     * @memberof Field
     */
    'settings'?: object;
    /**
     * For enum field types like SELECT or MULTI_SELECT
     * @type {Array<FieldOptionsInner>}
     * @memberof Field
     */
    'options'?: Array<FieldOptionsInner>;
}

export const FieldTypeEnum = {
    Uuid: 'UUID',
    Text: 'TEXT',
    Phones: 'PHONES',
    Emails: 'EMAILS',
    DateTime: 'DATE_TIME',
    Date: 'DATE',
    Boolean: 'BOOLEAN',
    Number: 'NUMBER',
    Numeric: 'NUMERIC',
    Links: 'LINKS',
    Currency: 'CURRENCY',
    FullName: 'FULL_NAME',
    Rating: 'RATING',
    Select: 'SELECT',
    MultiSelect: 'MULTI_SELECT',
    Relation: 'RELATION',
    Position: 'POSITION',
    Address: 'ADDRESS',
    RawJson: 'RAW_JSON',
    RichText: 'RICH_TEXT',
    RichTextV2: 'RICH_TEXT_V2',
    Actor: 'ACTOR',
    Array: 'ARRAY',
    TsVector: 'TS_VECTOR'
} as const;

export type FieldTypeEnum = typeof FieldTypeEnum[keyof typeof FieldTypeEnum];