// packages/backend-lib/src/twentyCrm/metaClient/model/field-for-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Twenty Api
 * This is a **Twenty REST/API** playground based on the **OpenAPI 3.1 specification**.
 *
 * The version of the OpenAPI document: v0.1
 * Contact: felix@twenty.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FieldOptionsInner } from './field-options-inner';

/**
 * A field
 * @export
 * @interface FieldForResponse
 */
export interface FieldForResponse {
    /**
     * 
     * @type {string}
     * @memberof FieldForResponse
     */
    'type'?: FieldForResponseTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof FieldForResponse
     */
    'objectMetadataId'?: string;
    /**
     * 
     * @type {string}
     * @memberof FieldForResponse
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof FieldForResponse
     */
    'label'?: string;
    /**
     * 
     * @type {string}
     * @memberof FieldForResponse
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof FieldForResponse
     */
    'icon'?: string;
    /**
     * 
     * @type {any}
     * @memberof FieldForResponse
     */
    'defaultValue'?: any;
    /**
     * 
     * @type {boolean}
     * @memberof FieldForResponse
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {object}
     * @memberof FieldForResponse
     */
    'settings'?: object;
    /**
     * For enum field types like SELECT or MULTI_SELECT
     * @type {Array<FieldOptionsInner>}
     * @memberof FieldForResponse
     */
    'options'?: Array<FieldOptionsInner>;
    /**
     * 
     * @type {string}
     * @memberof FieldForResponse
     */
    'id'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof FieldForResponse
     */
    'isCustom'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof FieldForResponse
     */
    'isActive'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof FieldForResponse
     */
    'isSystem'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof FieldForResponse
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof FieldForResponse
     */
    'updatedAt'?: string;
}

export const FieldForResponseTypeEnum = {
    Uuid: 'UUID',
    Text: 'TEXT',
    Phones: 'PHONES',
    Emails: 'EMAILS',
    DateTime: 'DATE_TIME',
    Date: 'DATE',
    Boolean: 'BOOLEAN',
    Number: 'NUMBER',
    Numeric: 'NUMERIC',
    Links: 'LINKS',
    Currency: 'CURRENCY',
    FullName: 'FULL_NAME',
    Rating: 'RATING',
    Select: 'SELECT',
    MultiSelect: 'MULTI_SELECT',
    Relation: 'RELATION',
    Position: 'POSITION',
    Address: 'ADDRESS',
    RawJson: 'RAW_JSON',
    RichText: 'RICH_TEXT',
    RichTextV2: 'RICH_TEXT_V2',
    Actor: 'ACTOR',
    Array: 'ARRAY',
    TsVector: 'TS_VECTOR'
} as const;

export type FieldForResponseTypeEnum = typeof FieldForResponseTypeEnum[keyof typeof FieldForResponseTypeEnum];