// packages/backend-lib/src/twentyCrm/metaClient/model/model-object.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Twenty Api
 * This is a **Twenty REST/API** playground based on the **OpenAPI 3.1 specification**.
 *
 * The version of the OpenAPI document: v0.1
 * Contact: felix@twenty.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * An object
 * @export
 * @interface ModelObject
 */
export interface ModelObject {
    /**
     * 
     * @type {string}
     * @memberof ModelObject
     */
    'nameSingular'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelObject
     */
    'namePlural'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelObject
     */
    'labelSingular'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelObject
     */
    'labelPlural'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelObject
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelObject
     */
    'icon'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelObject
     */
    'labelIdentifierFieldMetadataId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelObject
     */
    'imageIdentifierFieldMetadataId'?: string;
}