// packages/backend-lib/src/twentyCrm/metaClient/model/object-for-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Twenty Api
 * This is a **Twenty REST/API** playground based on the **OpenAPI 3.1 specification**.
 *
 * The version of the OpenAPI document: v0.1
 * Contact: felix@twenty.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ObjectForResponseFields } from './object-for-response-fields';

/**
 * An object
 * @export
 * @interface ObjectForResponse
 */
export interface ObjectForResponse {
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'nameSingular'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'namePlural'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'labelSingular'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'labelPlural'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'icon'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'labelIdentifierFieldMetadataId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'imageIdentifierFieldMetadataId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'dataSourceId'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ObjectForResponse
     */
    'isCustom'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ObjectForResponse
     */
    'isActive'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ObjectForResponse
     */
    'isSystem'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof ObjectForResponse
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {ObjectForResponseFields}
     * @memberof ObjectForResponse
     */
    'fields'?: ObjectForResponseFields;
}